Add attr module
